Move methods to data.utils
